move the input into a folder
