+++ src/common/components/forms/index.ts
@@ -1,4 +1,4 @@
-export * from './input-formik.component';
+export * from './input-formik';
 export * from './dropdown-formik';
 export * from './table-pk-picker-formik';
 export * from './checkbox-formik';
