+++ src/common/components/forms/input-formik/index.ts
@@ -1,0 +1,1 @@
+export * from './input-formik.component';
